+++ Jegarn/Packet/Text/JegarnTextGroupChatPacket.m
@@ -16,7 +16,7 @@
     dict[@"type"] = self.type;
     dict[@"content"] = @{
             @"type" : [JegarnTextGroupChatPacket packetSubType],
-            @"group_id" : [NSString stringWithFormat:@"%i", self.content.groupId],
+            @"group_id" : [NSString stringWithFormat:@"%i", (int) self.content.groupId],
             @"text" : self.content.text ? self.content.text : @"",
     };
